requirements/vllm-conda-list.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.10.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.10.0=pypi_0
astor=0.8.1=pypi_0
attrs=25.3.0=pypi_0
blake3=1.0.5=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.15=h06a4308_0
cachetools=6.2.0=pypi_0
cbor2=5.7.0=pypi_0
certifi=2025.8.3=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.3=pypi_0
click=8.2.1=pypi_0
cloudpickle=3.1.1=pypi_0
compressed-tensors=0.10.2=pypi_0
cupy-cuda12x=13.6.0=pypi_0
depyf=0.19.0=pypi_0
dill=0.4.0=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
einops=0.8.1=pypi_0
email-validator=2.3.0=pypi_0
expat=2.7.1=h6a678d5_0
fastapi=0.116.1=pypi_0
fastapi-cli=0.0.10=pypi_0
fastapi-cloud-cli=0.1.5=pypi_0
fastrlock=0.8.3=pypi_0
filelock=3.19.1=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.7.0=pypi_0
gguf=0.17.1=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.1.9=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.34.4=pypi_0
idna=3.10=pypi_0
interegular=0.3.3=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
llguidance=0.7.30=pypi_0
llvmlite=0.44.0=pypi_0
lm-format-enforcer=0.10.12=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
mistral-common=1.8.4=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.1=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.6.4=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.5=pypi_0
ninja=1.13.0=pypi_0
numba=0.61.2=pypi_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.8.3.14=pypi_0
nvidia-cuda-cupti-cu12=12.8.57=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.61=pypi_0
nvidia-cuda-runtime-cu12=12.8.57=pypi_0
nvidia-cudnn-cu12=9.7.1.26=pypi_0
nvidia-cufft-cu12=11.3.3.41=pypi_0
nvidia-cufile-cu12=1.13.0.11=pypi_0
nvidia-curand-cu12=10.3.9.55=pypi_0
nvidia-cusolver-cu12=11.7.2.55=pypi_0
nvidia-cusparse-cu12=12.5.7.53=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.8.61=pypi_0
nvidia-nvtx-cu12=12.8.55=pypi_0
openai=1.102.0=pypi_0
openai-harmony=0.0.4=pypi_0
opencv-python-headless=4.12.0.88=pypi_0
openssl=3.0.17=h5eee18b_0
outlines-core=0.2.10=pypi_0
packaging=25.0=pypi_0
partial-json-parser=0.2.1.1.post6=pypi_0
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
prometheus-client=0.22.1=pypi_0
prometheus-fastapi-instrumentator=7.1.0=pypi_0
propcache=0.3.2=pypi_0
protobuf=6.32.0=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
py-cpuinfo=9.0.0=pypi_0
pybase64=1.4.2=pypi_0
pycountry=24.6.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pydantic-extra-types=2.10.5=pypi_0
pygments=2.19.2=pypi_0
python=3.12.11=h22baa00_0
python-dotenv=1.1.1=pypi_0
python-json-logger=3.3.0=pypi_0
python-multipart=0.0.20=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=27.0.2=pypi_0
ray=2.49.0=pypi_0
readline=8.3=hc2a1206_0
referencing=0.36.2=pypi_0
regex=2025.8.29=pypi_0
requests=2.32.5=pypi_0
rich=14.1.0=pypi_0
rich-toolkit=0.15.0=pypi_0
rignore=0.6.4=pypi_0
rpds-py=0.27.1=pypi_0
safetensors=0.6.2=pypi_0
scipy=1.16.1=pypi_0
sentencepiece=0.2.1=pypi_0
sentry-sdk=2.35.1=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=78.1.1=py312h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.50.2=hb25bd0a_1
starlette=0.47.3=pypi_0
sympy=1.14.0=pypi_0
tiktoken=0.11.0=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.22.0=pypi_0
torch=2.7.1+cu128=pypi_0
torchaudio=2.7.1+cu128=pypi_0
torchvision=0.22.1+cu128=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.56.0=pypi_0
triton=3.3.1=pypi_0
typer=0.17.3=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=pypi_0
uvicorn=0.35.0=pypi_0
uvloop=0.21.0=pypi_0
vllm=0.10.1.1=pypi_0
watchfiles=1.1.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py312h06a4308_0
xformers=0.0.31=pypi_0
xgrammar=0.1.21=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
